Add more file existance checks in the main command loop

+++ src/main.cpp
@@ -28,23 +28,47 @@ int main() {
         } else if (cmd == "create" || cmd == "touch") {
             string filename;
             cin >> filename;
-            cout << (myfs::create(filename) ? "File created" : "File wasn't created") << endl;
+            if (!myfs::file_exists(filename)) {
+                cout << (myfs::create(filename) ? "File created" : "File wasn't created") << endl;
+            } else {
+                cout << "File already exists!" << endl;
+            }
         } else if (cmd == "link" || cmd == "ln") {
             string target, name;
             cin >> target >> name;
-            cout << (myfs::link(target, name) ? "Link created" : "Link wasn't created") << endl;
+            if (!myfs::file_exists(target)) {
+                cout << "Target file doesn't exist" << endl;
+            } else if (myfs::file_exists(name)) {
+                cout << "File with name '" << name << "' already exists" << endl;
+            } else {
+                cout << (myfs::link(target, name) ? "Link created" : "Link wasn't created") << endl;
+            }
         } else if (cmd == "unlink" || cmd == "rm") {
             string filename;
             cin >> filename;
-            cout << (myfs::unlink(filename) ? "Hard link was removed" : "Hard link wasn't removed") << endl;
+            if (!myfs::file_exists(filename)) {
+                cout << "File doesn't exist" << endl;
+            } else if (myfs::File(filename).type() == myfs::FileType::Directory) {
+                cout << "Cannot remove directory, use `rmdir` command" << endl;
+            } else {
+                cout << (myfs::unlink(filename) ? "Hard link was removed" : "Hard link wasn't removed") << endl;
+            }
         } else if (cmd == "mkdir") {
             string dirname;
             cin >> dirname;
-            cout << (myfs::mkdir(dirname) ? "Dir created" : "Dir wasn't created") << endl;
+            if (myfs::file_exists(dirname)) {
+                cout << "File with name '" << dirname << "' already exists" << endl;
+            } else {
+                cout << (myfs::mkdir(dirname) ? "Dir created" : "Dir wasn't created") << endl;
+            }
         } else if (cmd == "rmdir") {
             string dirname;
             cin >> dirname;
-            cout << (myfs::rmdir(dirname) ? "Dir successfully removed" : "Dir wasn't removed") << endl;
+            if (myfs::file_exists(dirname)) {
+                cout << (myfs::rmdir(dirname) ? "Dir successfully removed" : "Dir wasn't removed") << endl;
+            } else {
+                cout << "Directory doesn't exist";
+            }
         } else if (cmd == "cd") {
             string dirname;
             cin >> dirname;
@@ -54,16 +78,31 @@ int main() {
         } else if (cmd == "symlink") {
             string target, name;
             cin >> target >> name;
-            cout << (myfs::symlink(name, target) ? "Symlink created" : "Symlink wasn't created") << endl;
+
+            if (!myfs::file_exists(target)) {
+                cout << "Target file doesn't exist" << endl;
+            } else if (myfs::file_exists(name)) {
+                cout << "File with name '" << name << "' already exists" << endl;
+            } else {
+                cout << (myfs::symlink(name, target) ? "Symlink created" : "Symlink wasn't created") << endl;
+            }
         } else if (cmd == "filestat" || cmd == "stat") {
             string filename;
             cin >> filename;
-            myfs::File f{filename};
-            cout << f.filestat();
+            if (myfs::file_exists(filename)) {
+                myfs::File f{filename};
+                cout << f.filestat();
+            } else {
+                cout << "File with name '" << filename << "' doesn't exist" << endl;
+            }
         } else if (cmd == "read" || cmd == "cat") {
             string filename;
             cin >> filename;
-            cout << myfs::File{filename}.cat() << endl;
+            if (myfs::file_exists(filename)) {
+                cout << myfs::File{filename}.cat() << endl;
+            } else {
+                cout << "File with name '" << filename << "' doesn't exist" << endl;
+            }
         } else if (cmd == "write") {
             string filename;
             cin >> filename;
@@ -75,18 +114,26 @@ int main() {
                 data += s;
                 data += ' ';
             }
-            myfs::File f{filename};
-            f.truncate(data.size());
-            f.write(data.data(), data.size(), 0);
-            cout << "Data successfully written" << endl;
+            if (myfs::file_exists(filename)) {
+                myfs::File f{filename};
+                f.truncate(data.size());
+                f.write(data.data(), data.size(), 0);
+                cout << "Data successfully written" << endl;
+            } else {
+                cout << "File with name '" << filename << "' doesn't exist" << endl;
+            }
         } else if (cmd == "truncate") {
             string filename;
             int size;
             cin >> filename;
             cin >> size;
             if (cin.fail()) continue;
-            myfs::File f{filename};
-            cout << (f.truncate(size) ? "File was truncated" : "File wasn't trucated") << endl;
+            if (myfs::file_exists(filename)) {
+                myfs::File f{filename};
+                cout << (f.truncate(size) ? "File was truncated" : "File wasn't trucated") << endl;
+            } else {
+                cout << "File with name '" << filename << "' doesn't exist" << endl;
+            }
         } else {
             cout << "Uknown command!" << endl;
         }
